test: remove hard coded reportDefinition id
