+++ integration/host_vulnerability_test.go
@@ -43,17 +43,6 @@ var (
 	hostname  = "circle-ci-test-node"
 )
 
-// Configure the correct machine id for vuln host tests
-func init() {
-	lw, err := laceworkIntegrationTestClient()
-	if err == nil {
-		// Check if the default machineID(30) exists
-		if !machineIDExists(lw) {
-			setHostVulnTestMachineID(lw)
-		}
-	}
-}
-
 func TestHostVulnerabilityCommandListCves(t *testing.T) {
 	out, err, exitcode := LaceworkCLIWithTOMLConfig("vulnerability", "host", "list-cves")
 	expectedHeaders := []string{
@@ -418,6 +407,15 @@ func machineIDExists(lw *api.Client) bool {
 		return false
 	}
 	return resp.Ok
+}
+
+func setReportDefinition(lw *api.Client) {
+	resp, err := lw.V2.ReportDefinitions.List()
+	if err != nil {
+		return
+	}
+
+	testReportDefinitionID = resp.Data[0].ReportDefinitionGuid
 }
 
 func setHostVulnTestMachineID(lw *api.Client) {
